Official Version 1.0
This is the first major release of ElectronVisualized. It contains various improvements such as the utilization of DOCKER for efficient and fast server-side communication, effectively setting up both React and Flask applications in one repo, all in one Docker image. It also enabled me to deploy C-based dependencies on the server-side computer, which was intially deemed to be impossible due to lack of support from the developers of GPAW library and its nature of supporting local rendering only. I have managed, however, to import it on a web-based server, and run it fully on the browser in combination with the Heroku server, which is quite an achievement for sure.